+++ x/tokenfactory/keeper/before_send_test.go
@@ -124,3 +124,78 @@ func (s *KeeperTestSuite) TestBeforeSendHook() {
 		})
 	}
 }
+
+// TestInfiniteTrackBeforeSend tests gas metering with infinite loop contract
+// to properly test if we are gas metering trackBeforeSend properly.
+func (s *KeeperTestSuite) TestInfiniteTrackBeforeSend() {
+	s.SkipIfWSL()
+
+	for _, tc := range []struct {
+		name            string
+		wasmFile        string
+		tokenToSend     sdk.Coins
+		useFactoryDenom bool
+		expectedError   bool
+	}{
+		{
+			name:            "sending tokenfactory denom from module to module with infinite contract should panic",
+			wasmFile:        "./testdata/infinite_track_beforesend.wasm",
+			useFactoryDenom: true,
+		},
+		{
+			name:            "sending non-tokenfactory denom from module to module with infinite contract should not panic",
+			wasmFile:        "./testdata/infinite_track_beforesend.wasm",
+			tokenToSend:     sdk.NewCoins(sdk.NewInt64Coin("foo", 1000000)),
+			useFactoryDenom: false,
+		},
+		{
+			name:            "Try using no 100 ",
+			wasmFile:        "./testdata/no100.wasm",
+			useFactoryDenom: true,
+		},
+	} {
+		s.Run(fmt.Sprintf("Case %s", tc.name), func() {
+			// setup test
+			s.SetupTest()
+
+			// load wasm file
+			wasmCode, err := os.ReadFile(tc.wasmFile)
+			s.Require().NoError(err)
+
+			// instantiate wasm code
+			codeID, _, err := s.contractKeeper.Create(s.Ctx, s.TestAccs[0], wasmCode, nil)
+			s.Require().NoError(err, "test: %v", tc.name)
+			cosmwasmAddress, _, err := s.contractKeeper.Instantiate(s.Ctx, codeID, s.TestAccs[0], s.TestAccs[0], []byte("{}"), "", sdk.NewCoins())
+			s.Require().NoError(err, "test: %v", tc.name)
+
+			// create new denom
+			res, err := s.msgServer.CreateDenom(sdk.WrapSDKContext(s.Ctx), types.NewMsgCreateDenom(s.TestAccs[0].String(), "bitcoin"))
+			s.Require().NoError(err, "test: %v", tc.name)
+			factoryDenom := res.GetNewTokenDenom()
+
+			var tokenToSend sdk.Coins
+			if tc.useFactoryDenom {
+				tokenToSend = sdk.NewCoins(sdk.NewInt64Coin(factoryDenom, 100))
+			} else {
+				tokenToSend = tc.tokenToSend
+			}
+
+			// send the mint module tokenToSend
+			s.FundModuleAcc("mint", tokenToSend)
+
+			// set beforesend hook to the new denom
+			// we register infinite loop contract here to test if we are gas metering properly
+			_, err = s.msgServer.SetBeforeSendHook(sdk.WrapSDKContext(s.Ctx), types.NewMsgSetBeforeSendHook(s.TestAccs[0].String(), factoryDenom, cosmwasmAddress.String()))
+			s.Require().NoError(err, "test: %v", tc.name)
+
+			// track before send suppresses in any case, thus we expect no error
+			err = s.App.BankKeeper.SendCoinsFromModuleToModule(s.Ctx, "mint", "distribution", tokenToSend)
+			s.Require().NoError(err)
+
+			// send should happen regardless of trackBeforeSend results
+			distributionModuleAddress := s.App.AccountKeeper.GetModuleAddress("distribution")
+			distributionModuleBalances := s.App.BankKeeper.GetAllBalances(s.Ctx, distributionModuleAddress)
+			s.Require().True(distributionModuleBalances.IsEqual(tokenToSend))
+		})
+	}
+}